requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
faiss-cpu
rank-bm25
sentence-transformers
transformers
accelerate
sentencepiece
tqdm
pyarrow
